+++ stripe/api_resources/__init__.py
@@ -5,6 +5,7 @@ from __future__ import absolute_import, division, print_function
 
 from stripe.api_resources.error_object import ErrorObject, OAuthErrorObject
 from stripe.api_resources.list_object import ListObject
+from stripe.api_resources.search_result_object import SearchResultObject
 
 from stripe.api_resources import billing_portal
 from stripe.api_resources import checkout
@@ -14,6 +15,7 @@ from stripe.api_resources import radar
 from stripe.api_resources import reporting
 from stripe.api_resources import sigma
 from stripe.api_resources import terminal
+from stripe.api_resources import test_helpers
 
 from stripe.api_resources.account import Account
 from stripe.api_resources.account_link import AccountLink