+++ stripe/object_classes.py
@@ -7,6 +7,7 @@ from stripe import api_resources
 OBJECT_CLASSES = {
     # data structures
     api_resources.ListObject.OBJECT_NAME: api_resources.ListObject,
+    api_resources.SearchResultObject.OBJECT_NAME: api_resources.SearchResultObject,
     # business objects
     api_resources.Account.OBJECT_NAME: api_resources.Account,
     api_resources.AccountLink.OBJECT_NAME: api_resources.AccountLink,
@@ -91,6 +92,7 @@ OBJECT_CLASSES = {
     api_resources.terminal.ConnectionToken.OBJECT_NAME: api_resources.terminal.ConnectionToken,
     api_resources.terminal.Location.OBJECT_NAME: api_resources.terminal.Location,
     api_resources.terminal.Reader.OBJECT_NAME: api_resources.terminal.Reader,
+    api_resources.test_helpers.TestClock.OBJECT_NAME: api_resources.test_helpers.TestClock,
     api_resources.ThreeDSecure.OBJECT_NAME: api_resources.ThreeDSecure,
     api_resources.Token.OBJECT_NAME: api_resources.Token,
     api_resources.Topup.OBJECT_NAME: api_resources.Topup,