+++ tests/test_generated_examples.py
@@ -122,8 +122,7 @@ class TestGeneratedExamples(object):
 
     def test_filelink_update(self, request_mock):
         stripe.FileLink.modify(
-            "link_xxxxxxxxxxxxx",
-            metadata={"order_id": "6735"},
+            "link_xxxxxxxxxxxxx", metadata={"order_id": "6735"}
         )
         request_mock.assert_requested(
             "post", "/v1/file_links/link_xxxxxxxxxxxxx"
@@ -504,8 +503,7 @@ class TestGeneratedExamples(object):
 
     def test_creditnote_update(self, request_mock):
         stripe.CreditNote.modify(
-            "cn_xxxxxxxxxxxxx",
-            metadata={"order_id": "6735"},
+            "cn_xxxxxxxxxxxxx", metadata={"order_id": "6735"}
         )
         request_mock.assert_requested(
             "post", "/v1/credit_notes/cn_xxxxxxxxxxxxx"
@@ -663,8 +661,7 @@ class TestGeneratedExamples(object):
 
     def test_invoiceitem_update(self, request_mock):
         stripe.InvoiceItem.modify(
-            "ii_xxxxxxxxxxxxx",
-            metadata={"order_id": "6735"},
+            "ii_xxxxxxxxxxxxx", metadata={"order_id": "6735"}
         )
         request_mock.assert_requested(
             "post", "/v1/invoiceitems/ii_xxxxxxxxxxxxx"
@@ -1638,4 +1635,33 @@ class TestGeneratedExamples(object):
         request_mock.assert_requested(
             "post",
             "/v1/setup_intents/seti_xxxxxxxxxxxxx/verify_microdeposits",
+        )
+
+    def test_test_helpers_testclock_create(self, request_mock):
+        stripe.test_helpers.TestClock.create(frozen_time=123, name="cogsworth")
+        request_mock.assert_requested("post", "/v1/test_helpers/test_clocks")
+
+    def test_test_helpers_testclock_retrieve(self, request_mock):
+        stripe.test_helpers.TestClock.retrieve("clock_xyz")
+        request_mock.assert_requested(
+            "get",
+            "/v1/test_helpers/test_clocks/clock_xyz",
+        )
+
+    def test_test_helpers_testclock_list(self, request_mock):
+        stripe.test_helpers.TestClock.list()
+        request_mock.assert_requested("get", "/v1/test_helpers/test_clocks")
+
+    def test_test_helpers_testclock_delete(self, request_mock):
+        stripe.test_helpers.TestClock.delete("clock_xyz")
+        request_mock.assert_requested(
+            "delete",
+            "/v1/test_helpers/test_clocks/clock_xyz",
+        )
+
+    def test_test_helpers_testclock_advance(self, request_mock):
+        stripe.test_helpers.TestClock.advance("clock_xyz", frozen_time=142)
+        request_mock.assert_requested(
+            "post",
+            "/v1/test_helpers/test_clocks/clock_xyz/advance",
         )